엔티티 이동 parallel job으로 구현
필요: sampleunitComponentData에 destindex 추가

+++ Assets/1.Scripts/DOTS/Authoring(baker)/SampleUnitAuthoring.cs
@@ -6,7 +6,7 @@ using Unity.Mathematics;
 public class SampleUnitAuthoring : MonoBehaviour
 {
     
-    public int movementspeed;
+    public float movementspeed;
     public int hp;
     public int dmg;
     public class SampleUnitAuthoringBaker : Baker<SampleUnitAuthoring>
@@ -24,7 +24,7 @@ public struct SampleUnitComponentData : IComponentData
     public int order; //순서
     public int2 index; // 타일
     public int2 destIndex;
-    public int movementspeed; //이동 속도
+    public float movementspeed; //이동 속도
     public int hp; 
     public int dmg;
 }
